+++ crates/api/src/services/mod.rs
@@ -1,5 +1,7 @@
 //! This module contains global services.
 //!
+pub use self::clipboard::*;
 pub use self::settings::*;
 
+mod clipboard;
 mod settings;